requirements.txt
langchain==0.0.311
numpy==1.22.3
Pillow==9.5.0
Pillow==10.0.1
sentence_transformers==2.2.2
streamlit==1.27.0
